pyproject.toml
[tool.poetry]
name = "couveflow"
version = "0.1.0"
description = "Device manager for simple IoT systems for gardens"
authors = ["Thiago Sant Helena <21184066+shthiago@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
django-rest-framework = "^0.1.0"
ply = "^3.11"
pytest-django = "^4.5.2"
loguru = "^0.6.0"
django-factory_boy = "^1.0.0"
pytest-cov = "^4.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.15.5"
mypy = "^0.982"
bandit = "^1.7.4"
pytest = "^7.2.0"
pre-commit = "^2.20.0"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
